feat:add dto institution

+++ internal/dto/client/get.go
@@ -1,7 +1,7 @@
 package dto
 
 type GetClientReq struct {
-	Id int64 `validate:"required"`
+	Id int32 `validate:"required"`
 }
 
 type GetClientRes struct {
